requirements.txt
Flask==2.0.2
google-cloud-firestore==2.3.2
Flask-Cors==3.0.10
Werkzeug==2.2.2